Implement Role API, refactor server tests
