+++ src/api/exportable-lib/index.js
@@ -1,6 +1,7 @@
-import ClientFunctionBuilder from '../client-functions/client-function-builder';
-import SelectorBuilder from '../client-functions/selectors/selector-builder';
-import testControllerProxy from './test-controller/proxy';
+import ClientFunctionBuilder from '../../client-functions/client-function-builder';
+import SelectorBuilder from '../../client-functions/selectors/selector-builder';
+import RoleDescriptor from './role-descriptor';
+import testControllerProxy from '../test-controller/proxy';
 
 export default {
     ClientFunction (fn, options) {
@@ -13,6 +14,10 @@ export default {
         var builder = new SelectorBuilder(fn, options, { instantiation: 'Selector' });
 
         return builder.getFunction();
+    },
+
+    Role (loginPage, initFn) {
+        return new RoleDescriptor(loginPage, initFn);
     },
 
     t: testControllerProxy
