+++ test/server/data/test-suites/role-init-fn-is-not-a-function/testfile.js
@@ -1,0 +1,8 @@
+import { Role } from 'testcafe';
+
+fixture `Test`;
+
+Role('exampe.com', 123);
+
+test('yo', () => {
+});
