+++ test/server/data/test-suites/role-login-page-is-not-a-string/testfile.js
@@ -1,0 +1,8 @@
+import { Role } from 'testcafe';
+
+fixture `Test`;
+
+Role(123, () => {});
+
+test('yo', () => {
+});
